requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
mysqlclient==2.0.3
Pillow==7.0.0
whitenoise==6.2.0